pyproject.toml
[project]
name = "irispy-client"
version = "0.1.0"
description = "An Iris bot client in Python"
authors = [{ name = "dolidolih", email = "dolidolih@proton.me" }]
requires-python = ">=3.9"
readme = "README.md"
dependencies = [
    "requests",
    "websockets",
    "pillow",
    "httpx"
]

[project.scripts]
iris = "iris.cli:main"
